[MODIFY]solve git comment in dashboard
