+++ TaskManagement/accounts/services.py
@@ -16,10 +16,10 @@ class AccountManagement():
 
     def set_email(self, userdata, *args, **kwargs):
         """
-        store the data in EmailAddress model of created user using FK. 
+        store the data in EmailAddress model of created user using FK.
         """
         userd = EmailAddress.objects.get_or_create(user=userdata,
                                                     email=userdata.email,
-                                                    verified= True, 
-                                                    primary= True)
+                                                    verified=True,
+                                                    primary=True)
         return userd
