+++ TaskManagement/templates/active_projects.html
@@ -2,6 +2,7 @@
 {% load static %}
 {% load i18n %}
 {% load task_count_tags %}
+{% load projects_count_tags%}
 <title>{% block title %}Projects{% endblock %}</title>
 {% block content%}
 <div class="row">
