+++ TaskManagement/templates/dashboard.html
@@ -43,6 +43,7 @@
 {% url 'mytask_list' as mytask_list %}
 {% url 'project_tasklist' as project_tasklist %}
 
+
 <!-- Page Loader -->
 <div class="page-loader-wrapper">
     <div class="loader">
